CAMERA tool and configurables have been updated to take a CAMERA server list. Will connect to the first CAMERA server in the list and then successive servers are used as fall backs in the event a server dies.
CAMERA server now has the ability to manage its data directory. It will move to a new directory either after a set time, maximum number of files written or manually. CameraSwitch.exe and newDataDirectory.sh manage the manual switch.

Update to the configuration files for the servers and their start up scripts. The configuration file can now be passed as a command line option.

Added the scripts used in the LHCb online environment to start the servers as system processes.


git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@99485 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: f5bc6c2beb4bac5dcf1e1a630f1102f92d95bba2