changed polarity and scaleFactor in FieldCvs interface such that they obey following definition:
- polarity is the sign of the y-coordinate of the field in the center
of the magnet (more exactly at (0,0,5200)

- scaleFactor is the absolute value of the scale of the magnetic field
relative to some reference. In this case it is just the absolute value
of the scale with which the input field map is multiplied. Since this
is still ill defined, we should either remove this function, or make
the fieldSvc aware of what the nominal field is.


git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@38058 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: e0494e72ad20d755c0349e092725cbeaf49dd638