Added a generic Makefile for CMake projects.
The default LHCb Makefile has been replaced with a small one that
chooses the CMT or the CMake version.
The CMT version is used by default, while the CMake one can be enabled
with the variable USE_CMAKE (environment or make argument) if the
project contains a CMakeLists.txt.

From: Marco Clemencic <marco.clemencic@cern.ch>


git-svn-id: http://svn.cern.ch/guest/lhcb/LbScripts/trunk@147856 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: e7924ec52e026ab158243626baa9b466b3a583a7