Mass Uncertainty in TrackParticleRefitter
Modified function that retrieves uncertainty of mass in TrackParticleRefitter due to update in LorentzVectorWithError.

The function that was used to get the uncertainty on the mass from TrackParticleRefitter was
```
Gaudi::Math::LorentzVectorWithError::sigmamass()
```

This has been removed from the head of __LHCbMath__.
An alternative function has been identified in
```
Gaudi::Math::LorentzVectorWithError::m().error()
```
and now used.

Merge branch 'TPRefitterMass' of LHCb-SVN-mirrors/Alignment

See merge request !1


git-svn-id: http://svn.cern.ch/guest/lhcb/Alignment/trunk@202637 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: e0696e3421e89940d00a1f4a950d93fb898f7560